+++ changelog.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file (keepachangelog.com).
 
+## 0.3.0 - 2016-01-05
+### Added
+- Add value validation `check` property on setters.
+- Add `.bos`, `eos`, `.$`, `.reset`, `.at` methods.
+
 ## 0.2.0 - 2016-01-04
 ### Added
 - Support `readonly` properties.